move text to core
